+++ src/main/java/subway/auth/infrastructure/LoginInterceptor.java
@@ -3,7 +3,7 @@ package subway.auth.infrastructure;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 import subway.auth.application.AuthService;
-import subway.auth.exception.InvalidTokenException;
+import subway.common.exception.InvalidTokenException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
